[WIP] split visualization.py
