+++ optuna/visualization/__init__.py
@@ -1,0 +1,6 @@
+from optuna.visualization.contour import plot_contour  # NOQA
+from optuna.visualization.intermediate_values import plot_intermediate_values  # NOQA
+from optuna.visualization.optimization_history import plot_optimization_history  # NOQA
+from optuna.visualization.parallel_coordinate import plot_parallel_coordinate  # NOQA
+from optuna.visualization.slice import plot_slice  # NOQA
+from optuna.visualization.utils import is_available  # NOQA
